Refactor achievements tab module
